Revert "compile"
This reverts commit 936800068d1e44957cad96fc552f5469990feb8c.

+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCTicker.java
@@ -54,7 +54,7 @@ public class CMCTicker {
         return rank;
     }
 
-    public void setRank(int rank) {
+    public void setRank(BigDecimal rank) {
         this.rank = rank;
     }
